Versions 2.* do not work since 3.0 exists
Installing 2.2 or 2.3 causes the installation of the Tvision\RackspaceCloudFilesStreamWrapper v 3.0.* (due to a wrong version constraint in the `composer.json`).

This completely breaks these two version causing the following fatal error:

```bash
Fatal error: Class 'Tvision\RackspaceCloudFilesStreamWrapper\Service\RackspaceApi' not found in [...]/app/cache/dev/appDevDebugProjectContainer.php
```

This pull request should fix version 2.2 (branch), but we need also to update the 2.3 tag (being a tag I have no idea if there's a way to submit a pull request for that). I believe the version 2.3 should require the 2.3 version of the stream wrapper library, so things should be updated properly for the two versions. I suggest to use the `~` operator and to delete and recreate the two tags whether possible.

In the meanwhile, if someone needs to install the version 2.2 or 2.3, he can use this workaround: force the version of the RackspaceCloudFilesStreamWrapper library in his own `composer.json` file:

example for 2.3
```json
        "tvision/rackspace-cloud-files-bundle": "~2.3",
        "tvision/rackspace-cloud-files-streamwrapper": "~2.3",
```